2.2.Webpacking our First Assets{Adding the Script & Links Tags}
Adding the Script & Links Tags

What this means is... we now have one JavaScript file that contains all the code we need and one CSS file that contains all the CSS! All we need to do is add them to our page!
Open up templates/base.html.twig. Let's keep the existing stylesheets for now and add a new one: <link rel="stylesheet" href=""> the asset() function and the public path: build/app.css. At the bottom, add the script tag with src="{{ asset('build/app.js') }}". Oh, make that app.js.
If you're not familiar with the asset() function, it's not doing anything important for us. Because the build/ directory is our document root, we're literally pointing to the public path.
Let's try it! Move over, refresh and... hello, weird blue background. And in the console... yes! There's the log!
We've only started to scratch the surface of the possibilities of Webpack. So if you're still wondering: "why is going through this build process so useful?". Stay tuned. Because next, we're going to talk about the require() and import statements and start organizing our code.