5.1.Importing External Libraries & Global Variables
We already added the app entry files to our base layout: the <script> tag and the <link> tag both live here:
 109 lines  templates/base.html.twig
<html lang="en">
    <head>
        {% block stylesheets %}
            {{ encore_entry_link_tags('app') }}
        {% endblock %}
    </head>
    <body>
        {% block javascripts %}
            {{ encore_entry_script_tags('app') }}
        {% endblock %}
    </body>
</html>

This means that any time we have some CSS or JavaScript that should be included on every page, we can put it in app.js.
Look down at the bottom
Ah... we have a few script tags for external files and some inline JavaScript. Shame on me! Let's refactor all of this into our new Encore-powered system.
The first thing we include is jQuery... which makes sense because we're using it below. Great! Get rid of it. Not surprisingly... this gives us a nice, big error:

$ is not defined