5.2.Importing External Libraries & Global Variables{yarn add jquery --dev}
Installing a Library (jQuery)

No worries! One of the most wondrous things about modern JavaScript is that we can install third-party libraries properly. I mean, with a package manager. Find your terminal and run:
yarn add jquery --dev

The --dev part isn't important. Technically we only need these files during the build process... they don't need to be included on production... which is why the --dev makes sense. But in 99% of the cases, it doesn't matter. We'll talk about production builds and deploying at the end of the tutorial.

And... that was painless! We now have jQuery in our app.