5.4.Importing External Libraries & Global Variables{Forcing a Global jQuery Variable}
Forcing a Global jQuery Variable

The ultimate solution is to refactor all of your code from your templates and un-Webpack-ified JavaScript files into Encore. But... if you're upgrading an existing site, phew! You probably have a ton of JavaScript that expects there to be global $ or jQuery variables. Moving all of that into Encore all at once... it's, uh... not very realistic.
So, if you really want a global variable, you can add one with global.$ = $
That global keyword is special to Webpack. Try it now: refresh! It works!
But... don't do this unless you have to. I'll remove it and add some comments to explain that this is useful for legacy code.
Let's properly finish this next by refactoring all our code into app.js, which will include installing two more libraries and our first jQuery plugin... It turns out that jQuery plugins are a special beast.