6.1.Bootstrap & the Curious Case of jQuery Plugins
The inline code in base.html.twig isn't working anymore because we've eliminated the $ global variable. Woo! To make it work, let's move all this code into app.js. Instead of global variables, we're importing $ and that's why it's called $ down here:
25 lines  assets/js/app.js
import $ from 'jquery';
$('.dropdown-toggle').dropdown();
$('.custom-file-input').on('change', function(event) {
    var inputFile = event.currentTarget;
    $(inputFile).parent()
        .find('.custom-file-label')
        .html(inputFile.files[0].name);
});
It's all just local variables.

Try it now. Ok, it sorta works. It logs... then explodes. The error has some Webpack stuff on it, but it ultimately says:

  dropdown is not a function

Click the app.js link. Ah, it's having trouble with the dropdown() function. That is one of the functions that Bootstrap adds to jQuery. And... it makes sense why it's missing: we're running all of our code here, and then including Bootstrap:
 99 lines  templates/base.html.twig
<html lang="en">
    <body>
        {% block javascripts %}
            {{ encore_entry_script_tags('app') }}
            <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-beta.2/js/bootstrap.min.js" integrity="sha384-alpBpkh1PFOepccYVYDB4do5UnbKysX5WZXm3XxPqe5iKTfUKjNkCk9SaVuEZflJ" crossorigin=anonymous></script>
        {% endblock %}
    </body>
</html>
It's simply not adding the function in time! Well actually, it's a bit more than that. Even if we moved this script tag up, it still wouldn't work. Why? Because when you include Bootstrap via a script tag, it expects jQuery to be a global variable... and that - wonderfully - doesn't exist anymore.

Let's do this properly.