6.3.Bootstrap & the Curious Case of jQuery Plugins{Installing Bootstrap: yarn add bootstrap --dev}
Then, find your terminal and run:

yarn add bootstrap --dev
Oh, and how did I know that the package name was bootstrap? Just because I cheated and searched for it before recording. Go to https://yarnpkg.com/ and search for "Bootstrap". 9.7 million downloads... in the last 30 days... that's probably the right one.
And... it's done! Oh, and there's a little notice:

  bootstrap has an unmet peer dependency popper.js

We'll come back to that in a minute.