6.4.Bootstrap & the Curious Case of jQuery Plugins{Importing jQuery Plugins}
Importing jQuery Plugins

Back in app.js installing Bootstrap isn't enough. On top, add import 'bootstrap'.
Nope, we don't need to say import $ from or anything like that. Bootstrap is a jQuery plugin and jQuery plugins are... super weird. They do not return a value. Instead, they modify jQuery and add functions to it. I'll add a note here because... it just looks strange: it's weird that adding this allows me to use the tooltip() function, for example.

How Bootstrap Finds jQuery

But wait a second. If Bootstrap modifies jQuery... internally, how does it get the jQuery object in order to do that? I mean, jQuery is no longer global: if we need it, we need to import it. Well... because Bootstrap is a well-written library, it does the exact same thing. It detects that it's in a Webpack environment and, instead of expecting there to be a global jQuery variable, it imports jquery, just like we are.
And, fun fact, when two different files import the same module, they get back the same, one instance of it - a lot like Symfony's container. We import jQuery and assign it to $. Then, a microsecond later, Bootstrap imports that same object and modifies it:

 26 lines  assets/js/app.js
import $ from 'jquery';
import 'bootstrap'; // adds functions to jQuery
By the time we get past line 11, the $ variable has the new tooltip() function.

But... you may have noticed that, while I was talking about how awesome this is all going to work... my build was failing!
This dependency was not found: popper.js in bootstrap.js
This is awesome! Bootstrap has two dependencies: jQuery but also another library called popper.js. Internally, it tries to import both of them. But, because this is not installed in our project, it fails. By the way, if you're wondering:
Why doesn't Bootstrap just list this as a dependency in its package.json so that it's automatically downloaded for us?
Excellent question! And that's exactly how we would do it in the PHP world. Short answer: Node dependencies are complicated, and so sometimes it will work like this, but sometimes it's a better idea for a library to force us to install its dependency manually. That's called a /peer/ dependency.