8.7.Importing 3rd Party CSS + Image Paths{Moving our CSS into Encore}
Moving our CSS into Encore

Ok one more link tag to go. Remove it! Then, open css/styles.css, copy all of this, delete that file, and, in app.css, highlight the blue background and paste!
That's a simple step so... it should work, right? Nope! Check out the build failure:

  Module not found: Can't resolve ../images/space-nav.jpg in our assets/css/ directory.
It doesn't show the exact file, but we only have one. Ah, here's the problem:

 246 lines  assets/css/app.css
/* NAVIGATION */
.navbar-bg {
    background: url('../images/space-nav.jpg');
}
PhpStorm is super angry about it too! This background image references ../images/, which was perfect when the code lived in the public/css/ directory. But when we moved it, we broke that path!
This is awesome! Instead of us silently not realizing we did this, we get a build error. Amazing! We can't break paths without Webpack screaming.