9.1.Sass & Overriding Bootstrap Variables
What if I want to use Sass instead of normal CSS, or maybe Less or Stylus? Normally, that takes some setup: you need to create a system that can compile all of your Sass files into CSS. But with Encore, we get this for free!
Rename app.css to app.scss. Of course, when we do that, the build fails because we need to update the import in app.js:

 26 lines  assets/js/app.js
// any CSS you require will output into a single css file (app.css in this case)
import '../css/app.scss';
But the build still fails. Go check out the error. Woh! That's awesome! It basically says:

  Hey! How are you? Great weather lately, right? Listen, it looks like you're trying to load a Sass file. That's super! To do that, enable the feature in Encore and install these libraries.

This is the philosophy of Encore: give you a really solid, but small-ish core, and then offer a ton of optional features.