9.4.Sass & Overriding Bootstrap Variables{Organizing into Partials}
Organizing into Partials

This next part is optional, but I want to get organized... instead of having one big file, create a new directory called layout/. And for this top stuff, create a file called _header.scss. Little-by-little, we're going to move all of this code into different files. Grab the first section and put it into header
We'll import the new files when we finish.
Next is the "advertisement" CSS. Create another folder called components/. And inside, a new _ad.scss file. I'll delete the header... then move the code there. Let's keep going! For the article stuff, create _articles.scss, and move the code. Then, _profile.scss, copy that code... and paste.
For the "Create Article" and "Article Show" sections, let's copy all of that and put it into _article.scss. And for the footer, inside layout/, create one more file there called _footer.scss and... move the footer code.
And finally, copy the sortable code, create another components partial called _sortable.scss and paste. Now we can import all of this with @import './layout/header' and @import './layout/footer'.
Notice: you don't need the _ or the .scss parts: that's a Sass thing. Let's add a few more imports for the components: ad, articles, profile and sortable.
Phew! That took some work, but I like the result! But, of course, Encore is here to ruin our party with a build failure:

  Cannot resolve ./images/space-nav.jpeg

We know that error!