9.6.Sass & Overriding Bootstrap Variables{Adding Variables}
Adding Variables

To celebrate that we're processing through Sass, let's at least use one of its features. Create a new directory called helper/ and a new file called _variables.scss.
At the top of _header.scss, we have a gray background color.
Just to prove we can do it, in _variables, create a new variable called $lightgray set to #efefee. And back in headers, reference that: $lightgray.
We even get auto-completion on that! As soon as we save, the build fails!

  Undefined variable: "$lightgray"