10.2.Copying Files{Controlling the copy Destination}
Controlling the copy Destination

Let's see... go back to the docs. Here it is: you can give it a destination... and this has a few wildcards in it, like [path], [name] and [ext]. Oh, but use this second one instead: it gives us built-in file versioning by including a hash of the contents in the filename.
Back in our config, paste that. Before we restart Encore, shouldn't we delete some of these old files... at least to get them out of the way and clean things up? Nope! Well, yes, but it's already happening. One other optional feature that we're using is called cleanOutputBeforeBuild().
This is responsible for emptying the build/ directory each time we build.

Ok, go restart Encore: Ctrl+C, then:
yarn watch
Let's go check it out! Beautiful! Everything now copies to images/ and includes a hash.