11.1.Page-Specific JS: Multiple Entries
On the article show page, if you check the console... it's an error!

  $ is undefined
Coming from article_show.js. This shouldn't be surprising. And not just because I seem to make a lot of mistakes. Open that template and go to the bottom. Ah, this brings in a js/article_show.js file:
 86 lines  templates/article/show.html.twig
{% block javascripts %}
   {{ parent() }}
   <script src="{{ asset('js/article_show.js') }}"></script>
{% endblock %}
Go find that: in public/, I'll close build/ and... there it is:
16 lines  public/js/article_show.js
$(document).ready(function() {
  $('.js-like-article').on('click', function(e) {
    e.preventDefault();
      var $link = $(e.currentTarget);
      $link.toggleClass('fa-heart-o').toggleClass('fa-heart');
      $.ajax({
        method: 'POST',
        url: $link.attr('href')
      }).done(function(data) {
        $('.js-like-article-count').html(data.hearts);
      })
  });
});
This contains some traditional JavaScript from a previous tutorial. The problem is that the global $ variable doesn't exist anymore. If you look closely on this page, you'll see that, at the bottom, we include the app.js file first and then article_show.js. And, of course, the app.js file does import jQuery:

 26 lines  assets/js/app.js
import $ from 'jquery';
But as we learned, this does not create a global variable and local variables in Webpack don't leak beyond the file they're defined in.
So... this file is broken. And that's fine because I want to refactor it anyways to go through Encore so that we can properly import the variable on top.
Before we do that, let's organize one tiny thing. In assets/js, create a new components/ directory. Move get_nice_messages.js into that... and because that breaks our build... update the import statement in app.js to point here