11.4.Page-Specific JS: Multiple Entries{Importing CSS}
Importing CSS

Because we haven't imported any CSS yet from article_show.js, we already saw that Webpack was smart enough to not output a CSS file. But! Open up _articles.scss. Part of this file is CSS for the article show page... which doesn't really need to be included on every page.
Let's copy all of this code, remove it, and, at the root of the css/ directory, create a new file called article_show.scss and... paste!
Both app.js and article_show.js are meant to import everything that's needed for the layout and for the article show page. app.scss and article_show.scss are kinda the same thing: they should import all the CSS that's needed for each spot.
At the top of article_show.scss, we don't strictly need to do this, but let's @import 'helper/variables to drive home the point that this is a standalone file that imports anything it needs.
Finally, back in article_show.js add import '../css/article_show.scss'.
Ok, check your terminal! Suddenly, gasp! Webpack is outputting an article_show.css file! And wow! You can really see code splitting in action! That vendors~app~article_show.js probably contains jQuery, because Webpack saw that it's used by both entries and so isolated it into its own file so it could be re-used.
Anyways, back in show.html.twig copy the javascripts block, paste, rename it to stylesheets and then change to encore_entry_link_tags().
That should do it! Move over, refresh and... cool! The page still looks good and the heart still works. If you inspect element on this page, in the head, we have two CSS files: app.css to power the layout and article_show.css to power this page.
At the bottom, we have 4 JavaScript files to power the two entrypoints. By the way, WebpackEncoreBundle is smart enough to not duplicate the vendors~app~article_show.js script tag just because both entries need it. Smart!
Next: we are close to having our whole app in Encore. Let's refactor a bunch more un-Webpack-ified code.